# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Apex API",
    "description": "API de Apex — pipeline d'ingestion, recherche à facettes et espace client de Studio Chicane.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/media": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Liste des médias",
        "operationId": "list_media_api_v1_media_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shooting_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shooting Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "batch_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Batch Id"
            }
          },
          {
            "name": "unattached",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Unattached"
            }
          },
          {
            "name": "quarantined",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Quarantined"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MediaSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media/{media_id}": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Fiche média complète",
        "operationId": "get_media_api_v1_media__media_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media/{media_id}/file/{variant}": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Flux binaire médié (thumb | preview | hd) — jamais d'URL présignée vers R2",
        "operationId": "get_media_file_api_v1_media__media_id__file__variant__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Media Id"
            }
          },
          {
            "name": "variant",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "thumb",
                "preview",
                "hd"
              ],
              "type": "string",
              "title": "Variant"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media/{media_id}/attach": {
      "post": {
        "tags": [
          "media"
        ],
        "summary": "Rattachement manuel depuis le bac « à rattacher »",
        "operationId": "attach_media_api_v1_media__media_id__attach_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Media Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaAttachRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media/{media_id}/engagements": {
      "post": {
        "tags": [
          "media"
        ],
        "summary": "Rattachement manuel à un engagement (J2, `source='human'`)",
        "operationId": "add_media_engagement_api_v1_media__media_id__engagements_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Media Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaEngagementAttachRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaEngagementOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media/{media_id}/engagements/{engagement_id}": {
      "delete": {
        "tags": [
          "media"
        ],
        "summary": "Retirer un rattachement (J2)",
        "operationId": "delete_media_engagement_api_v1_media__media_id__engagements__engagement_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Media Id"
            }
          },
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Engagement Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media/{media_id}/ocr": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Candidats OCR bruts du média (J2) — score et boîte, affichés dans l'UI",
        "operationId": "get_media_ocr_api_v1_media__media_id__ocr_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaOcrResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MediaAttachRequest": {
        "properties": {
          "shooting_id": {
            "type": "integer",
            "title": "Shooting Id"
          }
        },
        "type": "object",
        "required": [
          "shooting_id"
        ],
        "title": "MediaAttachRequest"
      },
      "MediaEngagementAttachRequest": {
        "properties": {
          "engagement_id": {
            "type": "integer",
            "title": "Engagement Id"
          }
        },
        "type": "object",
        "required": [
          "engagement_id"
        ],
        "title": "MediaEngagementAttachRequest",
        "description": "Corps de `POST /media/{id}/engagements` (J2, rattachement manuel `source='human'`)."
      },
      "MediaEngagementOut": {
        "properties": {
          "engagement_id": {
            "type": "integer",
            "title": "Engagement Id"
          },
          "car_number": {
            "type": "string",
            "title": "Car Number"
          },
          "source": {
            "type": "string",
            "enum": [
              "ocr",
              "human"
            ],
            "title": "Source"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "engagement_id",
          "car_number",
          "source",
          "confidence"
        ],
        "title": "MediaEngagementOut"
      },
...
      "MediaOcrResponse": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/OcrCandidateOut"
            },
            "type": "array",
            "title": "Candidates"
          }
        },
        "type": "object",
        "required": [
          "candidates"
        ],
        "title": "MediaOcrResponse"
      },
      "MediaOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "batch_id": {
            "type": "integer",
            "title": "Batch Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "byte_size": {
            "type": "integer",
            "title": "Byte Size"
          },
          "mime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "shot_at_exif": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shot At Exif"
          },
          "shot_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shot At"
          },
          "exif": {
            "$ref": "#/components/schemas/MediaExif"
          },
          "phash": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phash"
          },
          "sharpness": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharpness"
          },
          "series_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Id"
          },
          "is_series_representative": {
            "type": "boolean",
            "title": "Is Series Representative"
          },
          "duplicate_of_media_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Of Media Id"
          },
          "ingest_status": {
            "type": "string",
            "enum": [
              "uploaded",
              "processing",
              "ingested",
              "quarantined"
            ],
            "title": "Ingest Status"
          },
          "quarantine_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quarantine Reason"
          },
          "quarantine_detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quarantine Detail"
          },
          "attachment_status": {
            "type": "string",
            "enum": [
              "unattached",
              "shooting_attached",
              "engagement_attached",
              "pending_review",
              "inconsistent"
            ],
            "title": "Attachment Status"
          },
          "attachment_source": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "pipeline_time",
                  "pipeline_ocr",
                  "human"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachment Source"
          },
          "attachment_detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachment Detail"
          },
          "shooting_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shooting Id"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          },
          "engagements": {
            "items": {
              "$ref": "#/components/schemas/MediaEngagementOut"
            },
            "type": "array",
            "title": "Engagements"
          },
          "events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "id",
          "batch_id",
          "original_filename",
          "byte_size",
          "mime",
          "width",
          "height",
          "shot_at_exif",
          "shot_at",
          "exif",
          "phash",
          "sharpness",
          "series_id",
          "is_series_representative",
          "duplicate_of_media_id",
          "ingest_status",
          "quarantine_reason",
          "quarantine_detail",
          "attachment_status",
          "attachment_source",
          "attachment_detail",
          "shooting_id",
          "is_simulated",
          "caption",
          "keywords",
          "engagements",
          "events"
        ],
        "title": "MediaOut",
        "description": "Fiche média complète : EXIF, série, doublon, engagements, journal d'événements."
      },
      "MediaSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "thumb_url": {
            "type": "string",
            "title": "Thumb Url"
          },
          "shot_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shot At"
          },
          "ingest_status": {
            "type": "string",
            "enum": [
              "uploaded",
              "processing",
              "ingested",
              "quarantined"
            ],
            "title": "Ingest Status"
          },
          "attachment_status": {
            "type": "string",
            "enum": [
              "unattached",
              "shooting_attached",
              "engagement_attached",
              "pending_review",
              "inconsistent"
            ],
            "title": "Attachment Status"
          },
          "shooting_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shooting Id"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated"
          },
          "duplicate_of_media_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Of Media Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "thumb_url",
          "shot_at",
          "ingest_status",
          "attachment_status",
          "shooting_id",
          "is_simulated",
          "duplicate_of_media_id"
        ],
        "title": "MediaSummary"
      },
      "OcrCandidateOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "raw_text": {
            "type": "string",
            "title": "Raw Text"
          },
          "normalized_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Number"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "bbox": {
            "additionalProperties": true,
            "type": "object",
            "title": "Bbox"
          },
          "engine_version": {
            "type": "string",
            "title": "Engine Version"
          },
          "resolution": {
            "type": "string",
            "enum": [
              "auto",
              "review",
              "abstain",
              "not_engaged",
              "accepted",
              "rejected"
            ],
            "title": "Resolution"
          },
          "engagement_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engagement Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "raw_text",
          "normalized_number",
          "confidence",
          "bbox",
          "engine_version",
          "resolution",
          "engagement_id"
        ],
        "title": "OcrCandidateOut",
        "description": "Candidat brut persisté — score et boîte, affichés dans l'UI (`GET /media/{id}/ocr`)."
      },
...
      "Page_MediaSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MediaSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[MediaSummary]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}